moves src/app/utils to src/utils
