+++ src/app/patterns/explorer/ProbeExplorer.svelte
@@ -14,7 +14,7 @@ import { formatBuildIDToDateString } from '../utils/formatters';
 
 import { histogramSpring } from '../utils/animation';
 
-import { extractBinValues } from '../../utils/probe-utils';
+import { extractBinValues } from '../../../utils/probe-utils';
 
 export let data;
 export let title;
